feat(alerts): extract shared alert sweep scheduling helpers
Both alerting sweeps (logs, billing) discover due alerts with the same
predicate and advance next_check_at with the same cadence math. Extract the
pure scheduling math (grid-snapped advancement, per-alert shard offsets) into
common/alerting/scheduling.py and the due-alert Q predicate into
posthog/alerting/scheduling.py, with the products' snooze-exclusion
divergence expressed as a parameter.

The workflow structures stay product-owned on purpose: logs runs a minutely
cron with in-workflow cohort batching, billing an hourly cron with child
workflows — unifying those would rewrite Temporal histories for no behavior
gain. The shared surface is the vocabulary both use, ready for insight
alerts to adopt.

Generated-By: PostHog Code
Task-Id: 041f4d4e-2e41-4347-88a3-1c9e7e55751d
